[NO-TSK] Fix: wishCnt 올리는 부분 추가

+++ src/main/kotlin/com/wafflestudio/team2server/post/service/ProductPostServiceImpl.kt
@@ -147,6 +147,7 @@ class ProductPostServiceImpl(
 		)
 	}
 
+	@Transactional
 	override fun getPostById(id: Long, authUserInfo: AuthUserInfo?): ProductPost {
 		val userId = authUserInfo?.uid
 		val postEntity: ProductPostEntity = productPostRepository.findById(id).getOrNull() ?: throw PostNotFoundException
@@ -166,20 +167,20 @@ class ProductPostServiceImpl(
 	}
 
 	@Transactional
-	override fun likePost(userId: Long, postId: Long) {
-		if (!wishListRepository.existsByUserIdAndPostId(userId, postId)) {
+	override fun likePost(userId: Long, postId: Long, enable: Boolean) {
+		if (enable && !wishListRepository.existsByUserIdAndPostId(userId, postId)) {
+			val postEntity = productPostRepository.findById(postId).getOrNull() ?: throw PostNotFoundException
 			val wishListEntity = WishListEntity(userId = userId, postId = postId)
+			postEntity.wishCnt++
 			wishListRepository.save(wishListEntity)
-		}
-	}
-
-	@Transactional
-	override fun unlikePost(userId: Long, postId: Long) {
-		if (wishListRepository.existsByUserIdAndPostId(userId, postId)) {
+		} else if (!enable && wishListRepository.existsByUserIdAndPostId(userId, postId)) {
+			val postEntity = productPostRepository.findById(postId).getOrNull() ?: throw PostNotFoundException
 			val wishListEntity = wishListRepository.findByUserIdAndPostId(userId = userId, postId = postId)
+			postEntity.wishCnt--
 			wishListRepository.delete(wishListEntity)
 		}
 	}
+
 
 	override fun getLikedPosts(userId: Long): List<PostSummary> {
 		return wishListRepository.findByUserId(userId).mapNotNull {
@@ -207,7 +208,7 @@ class ProductPostServiceImpl(
 	}
 
 	override fun getPostListRandom(
-		cur: Long,
+		prevCur: Long,
 		seed: Int,
 		distance: Int,
 		count: Int,
@@ -217,7 +218,7 @@ class ProductPostServiceImpl(
 		if (areaId !in authUserInfo.refAreaIds) {
 			throw BaniException(ErrorType.INVALID_PARAMETER)
 		}
-		val cur = if (count % 300 == 0) Long.MAX_VALUE else cur
+		val cur = if (count % 300 == 0) Long.MAX_VALUE else prevCur
 		val fetch = productPostRepository.findRandom(cur, seed, areaService.getAdjAreas(areaId, distance), (count / 300) * 300)
 		return ListResponse(
 			fetch.subList(0, min(15, fetch.size)).map {
@@ -348,6 +349,5 @@ class ProductPostServiceImpl(
 			status = post.status.name,
 		)
 	}
-
 }
